Add files via upload
Finished my solution for permutation, initially with a loop + recursion and then with 2 recursions and no loop. My solution is based on they way we do it combinatorically - choose the first character, it has n options if the string is of length n, then for the next characters there are n-1 options, and so on.