Add files via upload
Finished fixing school solution to my variation of findSum() problem - I do not allow to move to neighbour cells to use for the sum - only immediat cells. School solution take any connected neighbour, even if 3 or 4 cells awaya, as long as it connected to the original cell. That problem is much simpler than mine.